refactor: cleanup bottom sheet code; use edge insets
